backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
langchain-huggingface
langchain-google-genai
langchain-text-splitters
python-dotenv
